refactor: move files around and delete unused ones
