+++ src/controllers/Clients.py
@@ -3,10 +3,10 @@ from typing import Optional
 from argon2 import PasswordHasher
 from redis import asyncio as aioredis
 
-import modules.utils.Logger as Logger
-from modules.utils.Generators import Generators
-from modules.models.ClientModels import ClientModel
-from modules.utils.RedisConnector import RedisConnector
+import src.utils.Logger as Logger
+from src.utils.Generators import Generators
+from src.models.ClientModels import ClientModel
+from src.utils.RedisConnector import RedisConnector
 
 config: dict = toml.load("config.toml")
